feat: 推广汇报接真实数据源 — 广告API+真实订单+决策引擎商品源
日报/周报假数据根因修复（店铺 8 SKU 在售但汇报永远全 0）：
- ozon-ads.ts(新): Ozon Performance API 客户端（token 缓存/俄式数字兼容/env 凭据优雅降级）
- queryDailyStats/querySalesRanking 改查 ozon_orders 真实订单（含取消单统计、products_json 聚合）
- queryPromoCost 废弃 token_usage 当广告费：真实 CPC 花费+广告订单收入算 ROI，不可用返回 null
- /api/stats/weekly 删除硬编码 orders:0/byDay:[]，改真实分组聚合
- performance.ts: 广告无数据显示'未接入'而非 0.00x，建议模板不再空数据硬套
- inventory.route.ts: product_performance 空表时 fallback Ozon API 实时拉在售商品（list/info/prices/stocks 四端点串联，sku_1688_mapping 补成本，90s 缓存）
- decision-engine: 空评分不再发空卡片
- schema: local_orders/ozon_orders/purchase_1688 的 order_id 列 INTEGER→BIGINT（修 Ozon 11 位订单号写入溢出，8/10-8/11 sync 154 次 errors:1 的真凶）
- 测试 +18（ozon-ads 5 + promo-db 7 + inventory-fallback 7 + decision-engine 2 中含修正），全量 403 绿

+++ apps/api-services/src/db/schema.ts
@@ -48,7 +48,7 @@ export async function initSchema(db: DbAdapter): Promise<void> {
 
     CREATE TABLE IF NOT EXISTS local_orders (
       id TEXT PRIMARY KEY, store_id TEXT NOT NULL DEFAULT 'store_1',
-      posting_number TEXT UNIQUE NOT NULL, order_id INTEGER NOT NULL,
+      posting_number TEXT UNIQUE NOT NULL, order_id BIGINT NOT NULL,
       status TEXT NOT NULL, created_at TIMESTAMP NOT NULL,
       updated_at TIMESTAMP DEFAULT NOW(),
       buyer_name_masked TEXT, buyer_phone_masked TEXT,
@@ -335,7 +335,7 @@ export async function initSchema(db: DbAdapter): Promise<void> {
       id TEXT PRIMARY KEY,
       store_id TEXT NOT NULL DEFAULT 'store_1',
       posting_number TEXT NOT NULL,
-      order_id INTEGER NOT NULL,
+      order_id BIGINT NOT NULL,
       order_number TEXT,
       status TEXT NOT NULL,
       created_at_ozon TEXT,
@@ -363,7 +363,7 @@ export async function initSchema(db: DbAdapter): Promise<void> {
       id TEXT PRIMARY KEY,
       store_id TEXT NOT NULL DEFAULT 'store_1',
       ozon_posting_number TEXT NOT NULL,
-      ozon_order_id INTEGER NOT NULL,
+      ozon_order_id BIGINT NOT NULL,
       source_1688_url TEXT,
       offer_id TEXT,
       sku_list_json TEXT NOT NULL DEFAULT '[]',
